This commit is to increase the number of query chunks used for long queries. We have a fixed number of tasks (16), that will run in a loop (with yields) and pull chunks from a per-pipeline queue and process them. For short queries, we still have 16 chunks. Longer queries can have an unbounded number of chunks.
We also report incremental progress as chunks processing starting (earlier, we
only reported a change in process when the pipeline was started, and when the
final merge was completed.

As part of the change, we are also insuring that flow queries are merged
correctly; we were earlier assuming that the per-task merge output buffer was always empty.
This is not true when a single task processes multiple chunks.