Use absolute value for start_time/end_time in OverlayToUnderlayFlowMapper utility
Issue:
If the start_time/end_time in the OverlayToUnderlayFlowMap query is
specified as relative time, then the actual start_time/end_time for
the FlowRecordTable query and UFlowData query would be different.
Since the FlowRecordTable is queried first and the result of
which is used to query the UFlowData table, the result may
not be correct if the start_time/end_time is different for
FlowRecord and UFlowData queries.

Fix:
Convert the relative start_time/end_time in the OverlayToUnderlayFlowMap
query to absolute time and use the same start_time/end_time for
FlowRecordTable and UFlowData queries.

Change-Id: I0bf708a4ff073f57e5951bd94e2be306889407c0
Closes-Bug: #1403427