This commit adds stats for Alarmgen The "keys" attribute tracks UVEKeys for which alarm processing was done. The "notifs" attributes tracks UVE Notifications received from the Collectors
Also, the sequence of notification reads from Kafka is changed.
When starting a partition, we first read the latest message, and move the read pointer to that point.
Then we read all known UVE Keys for that partition from Redis.
Then we start reading all messages starting at the read pointer.

The sequence ensure that we do not miss any UVE change, while also allowing us to ignore
older UVE Changes that are already captured by the partition information in Redis.

Change-Id: I0e0ecbfad103274bcd19490a6dcb216e25a966ba
Partial-Bug:1428271