Fix analytics_systest failure
analytics_systest failed due to assert in verify_keyword_query().
This is a timing issue and therefore not seen consistently.
Presently, the retry logic doesn't kick in for failure case because
of the assert.
Removed the assert so that the evaluation of the return condition
triggers the retry logic in case of failure.

Change-Id: Iae29b0db3797800ca37844c60951fde2cc2bd49b
Closes-Bug: #1436231