New partitioning scheme for UVEAggregation: There 5 partitions ranges.
1: For all node UVE other than vrouter, server and pnode
2: For vrouter, server and pnode UVEs
3: For VM UVE
4: For Physical Interface, Logical Inteface and VMI UVEs
5: For everything else

Right now the default number of partitions is 1,2,4,4 and 4.
This can be changed in the conf files for collector, analytics-api and alarm-gen
This will make it more efficient to support UVE table-specific UVE streams

TODO:
a. support UVE table-specific UVE streams in analytics-api
b. raise alarm from analytics-api if the expected number of partitions in not up

Change-Id: I0ceda9d511ea22a1d98a593a46726dc57fa70e8d
Partial-Bug: 1428271