Implement alarm_ack_callback in alarm-generator
This patch implements alarm_ack_callback method in alarm-generator.
This method is passed as a parameter in the init_generator().
Upon receiving the SandeshAlarmAckRequest, the corresponding
handler defined in the sandesh library would invoke this callback.
The alarm_ack_callback, on successful validation of the ack request,
sends the sandesh alarm with 'ack' field set to True.

Change-Id: I7bcddf07f6b7cda189e904046207ae9e0c448239
Closes-Bug: #1512346