Escape special characters in the http query parameters
query parameters in the introspect request containing special
characters such as '&' need to be escaped by the http client.
Not escaping the special characters would result in improper
parsing of the http request at the server.
contrail-analytics-api uses the introspect_util module in the opserver
python package to send the introspect request to the generator.
Escape special characters in the query params using urllib.urlencode()

Change-Id: I843f571b0fe788d8ed161664880ed93cc3d00251
Closes-Bug: #1515489