Send DatabasePurgeInfo as objectlog instead of UVE
After database purge operation, analytics-api sends DatabasePurgeInfo
UVE. If database-node is not provisioned on the same node as analytics,
then a false alarm is raised due to missing NodeStatus and
ContrailConfig UVE.

DatabasePurgeInfo is not an ideal UVE candidate as upon restart of the
analytics-api service, old DatabasePurgeInfo UVE would not be sent
again. Therefore, converted DatabasePurgeInfo to objectlog.

Change-Id: I7e0308694e6c7bc55171fdcd87fd75dce05133e7
Closes-Bug: #1517273