Let Analytics services connect to any collector based on discovery
Presently, the analytics services such as analytics-api, query-engine,
snmp-collector and topology always connect to local collector. If the
local collector service is down, then these services do not connect to
other active collector service, if present.

Side-effect of this behavior:
If the collector service is down, then the PRouterEntry UVEs and the
PRouterLinkEntry UVEs originated by the snmp-collector and the topology
service respectively would be missing from the /analytics/uves/prouter
output.

Fix:
analytics-api, query-engine, snmp-collector and topology should get the
collector list from the discovery service instead of connecting to the
local collector service.

Change-Id: I62074e8516f4e7eb88ac1d1c727067e87578794f
Partial-Bug: #1528770