Alarmgen should ensure that we always able to catch the gevent kill exception in worker gevents. We go into a sleep after a kafka error. We should sleep in a section that can catch exceptions
AddressMismatchCompute and VrouterInterface alarms should not be raised when the VrouterAgent struct is absent.
AddressMismatchControl alarms should not be raised when the BgpRouterState struct is absent

When we check the object-type in api-server (during ContrailConfig insertion), we must replace "-" by "_"
Use "oper" instead of obj_dict to decide whether the object is being deleted. See:
https://github.com/Juniper/contrail-controller/commit/bf6cba2bb5b061bfd56fa99c72b7b2323f08486f

Change-Id: I2d3f8f66ac8ad8df13db816936e98ed7320c54a0
Closes-Bug: #1533158
Closes-Bug: #1536085