alarm rule based on stats results in continuous generation of alarms
For the alarm rule with operations >=, >, <=, <, range, the change in
operand value results in continuous generation of alarms. The
expectation is that the alarm should be raised only once.

This patch fixes this issue by ignoring the operand values if the
operation is any of [>=, >, <=, <, range] while comparing the existing
alarm with new alarm.

Change-Id: I8964626f439d4b873cf18b5fe36e820cec1e9ffc
Closes-Bug: #1616685
(cherry picked from commit 41e31bb35b576718377e31dd532396f9c8b99a73)