Configure kafka and kazoo loggers in contrail-alarm-gen
When contrail-alarm-gen disconnects from kafka, it retries
to reconnect to it and spits out bunch of messages. The kafka
python library logger was configured to output those messages
to stdout and due to this supervisor-analytics would get very
busy and not be able to handle other commands like stop all
and hence service supervisor-analytics stop would hang. Fix is
to configure the kafka and kazoo loggers in contrail-alarm-gen
as per the logging parameters and log to the file instead.

Change-Id: I553c53b3b549269a8cc0a87feda137980f008643
Closes-Bug: #1706832