Fix for analytics-api UVE query response delay
For every UVE request, analytics-api creates a connection to redis to
read the UVE. Creating a new redis connection for every UVE request is an
overhead and the response time depends on the gevent scheduling.
With this patch, analytics-api shares the same redis connections across the
UVE requests and therefore ensures the ordering of the UVE response per
redis, thereby alleviating the disproportionate response time for some
UVE requests.

Change-Id: I1066255d1a70218875ba6b61c29c6cd65c1a94c2
Closes-Bug: #1731187