Add timestamp to UVE content
At present, the timestamp field __T is added by alarm-gen for each UVE
struct when it aggregates the UVE. When use_aggregated_uve_db option is
set in the contrail-analytics-api.conf, then analytics-api doesn't rely
on alarm-gen for UVE aggregation and hence the existing design won't
work for this case.

This patch changes the model in which __T is added to the UVE structure.
collector will add __T for each UVE structure while writing in redis.
For cases when the same structure is sent by different sources, then the
latest timestamp  __T will be returned in the UVE response.

Closes-bug: 1776308

Change-Id: Iaf801ae385acd31bae6afbcf99cee474865b77de