Analytics queries are timing out when redis tls is enabled.
RCA - For query request, analytics-api server posts the query to localhost
redis server, which in turn notifies query-engine to process the query.
In case of TLS enabled, analytics-api server was connecting to localhost
redis server directly and passing the SSL flag as TRUE. This makes redis to go
unresponsive since it doesn't support SSL natively.

Fix - When analytics-api server connects to localhost redis directly, it should
pass SSL flag as false.

UT - In opserver UT, redis and stunnel both listen on 127.0.0.1 but on different
ports. In this case we have to disable redis tls for analytics-api as mentioned
in fix. So, I'm adding another parameter opserver_redis_ssl in AnalyticsFixture,
which overrides the redis_ssl_params, if it is set to False.

Change-Id: I3951fa389be989be1d81cb064431abaa5881f7b3
closes-jira-bug: CEM-7813