Query-engine is crashing if it connects to more than 5 redis instances
There is 2D array to manage redis connections which is nRedis x nConnections.
But while declaring the array, nConnections (=5) was made as Row of that 2D array.
So, when it connects to 6th redis instances, it will update 6th row of that 2D
array. This leads to crash due to invalid memory access.
We can't cherry pick from master as master code is in tf-analytics and 1912 code is
in contrail-analytics

closes-jira-bug: CEM-12908
Change-Id: I33f261a83c2ae94d4b4ed4eee6112523737bafbe