Cherry pick https://review.opencontrail.org/c/tungstenfabric/tf-analytics/+/56328
Contrail collector is hogging memory

It was observed that collector is continuosly hogging memory and making
the system unstable. There was no leak observed when collector was in
valgrind within UT. Also, core-analyzer utility didn't report any memory
leak. However it showed that collector is holding ~16 GB of free memory
and
it was not released back to system. This was causing memory hog.
In this fix, malloc_trim call is added to release the memory to system
periodically.
This fix is verified on customer setup and since this is called
periodically,
any memory hog will be temporary.

closes-jira-bug: CEM-21323
Change-Id: I783e9e036ef126a2591da05041efcfd2bd5702b8
(cherry picked from commit 92a2d44bfaa32330a17dc0df535c8e96e8d0f39e)