Collector node not created in zookeeper in case of connectivity failure
In case of, one node receives connectivity failure then collector
node in zookeeper gets deleted and when connectivity restores then
collector node doesn't gets created in zookeeper because listener
was not there to listen if connectivity restores or not.

So, When connectivity restores zookeeper make a callback function
stating that the earlier connnection was expired.

So, I have created a function with all parameters which zookeeper make
a callback and passed that function pointer to zookeeper_init so that
if zookeeper make a callback then all events should be handled properly.

closes-jira-bug: CEM-17031
Change-Id: I49df5cc2ecdab7219ba93d186307fa581efda9c1
Depends-On: I16e3d0873b50393aa15da508b3735055ec652ece
Depends-On: Id81c39386dd2c60ef1e6fc0f9f633dd7048c6e9a
Depends-On: Ifc63c436ce60357d067273c2d943a827193f5fd0